# ui/client/shared/api/openapi.json
{
  "components": {
    "schemas": {
      "AbsenceRequest": {
        "description": "Request body for recording an absence.\n\nOne per (project, date): a second one posted for the same date replaces\nthe first, so this is also how a half day becomes a full one.",
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "half_day": {
            "default": false,
            "title": "Half Day",
            "type": "boolean"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "type": {
            "$ref": "#/components/schemas/AbsenceType"
          }
        },
        "required": [
          "date",
          "type"
        ],
        "title": "AbsenceRequest",
        "type": "object"
      },
      "AbsenceResponse": {
        "description": "Response schema for an absence.",
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "half_day": {
            "default": false,
            "title": "Half Day",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AbsenceType"
          }
        },
        "required": [
          "id",
          "project_id",
          "date",
          "type"
        ],
        "title": "AbsenceResponse",
        "type": "object"
      },
      "AbsenceType": {
        "description": "Why a day was not worked. For the record and the calendar's colour only \u2014\nthe arithmetic treats all three the same.",
        "enum": [
          "vacation",
          "sick",
          "other"
        ],
        "title": "AbsenceType",
        "type": "string"
      },
...
      "Contract": {
        "description": "A day job's terms as they changed over time, plus what frames them.\n\n`terms` is the history \u2014 part-time, then 80%, then full-time \u2014 kept in\norder of `effective_from` with no two starting on the same day, because\nthe day decides which term applies. The region names the employer's\npublic holidays: `holiday_subdivision` only means something inside a\n`holiday_country`. Both must be codes the holidays library knows, but\nthat is checked where a contract is written (`ProjectService`), not\nhere: this model is re-validated on every read, and the set of codes the\nlibrary knows moves with its version \u2014 a code it stops recognising must\nmake one contract un-editable, not every project of the user unreadable.\n`opening_balance_hours` is what was banked (or owed) before Beats started\ncounting; `ended_on` freezes the balance from that day on.",
        "properties": {
          "ended_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended On"
          },
          "holiday_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holiday Country"
          },
          "holiday_subdivision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holiday Subdivision"
          },
          "opening_balance_hours": {
            "default": 0,
            "title": "Opening Balance Hours",
            "type": "number"
          },
          "terms": {
            "items": {
              "$ref": "#/components/schemas/ContractTerm"
            },
            "title": "Terms",
            "type": "array"
          }
        },
        "required": [
          "terms"
        ],
        "title": "Contract",
        "type": "object"
      },
...
      "CreateProjectRequest": {
        "description": "Request body for creating a project.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "contract": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Contract"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "kind": {
            "$ref": "#/components/schemas/ProjectKind",
            "default": "side_project"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "weekly_goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Goal"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
...
      "GoalOverrideResponse": {
        "description": "Response schema for a goal override (mirrors the request shape).",
        "properties": {
          "effective_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "goal_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoalType"
              },
              {
                "type": "null"
              }
            ]
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "week_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week Of"
          },
          "weekly_goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Goal"
          }
        },
        "title": "GoalOverrideResponse",
        "type": "object"
      },
      "GoalType": {
        "description": "Type of weekly goal: target to reach or cap to stay under.",
        "enum": [
          "target",
          "cap"
        ],
        "title": "GoalType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectKind": {
        "description": "What a project is to the person tracking it. A contract is read only on a day job.",
        "enum": [
          "day_job",
          "freelance",
          "side_project"
        ],
        "title": "ProjectKind",
        "type": "string"
      },
      "ProjectResponse": {
        "description": "Canonical response shape for a project \u2014 every field of the domain\nProject the API exposes. Previously declared only 6 of 11 fields;\nlist/update routes returned `model_dump()` with no response_model\ndeclared, so the OpenAPI contract was silently widened. Now precise so\ngenerated clients see the full shape.\n\n`contract` is the domain model as stored \u2014 the terms, region, opening\nbalance and end date \u2014 and is only meaningful when `kind` is `day_job`.",
        "properties": {
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "autostart_repos": {
            "items": {
              "type": "string"
            },
            "title": "Autostart Repos",
            "type": "array"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "contract": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Contract"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "github_repo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repo"
          },
          "goal_overrides": {
            "items": {
              "$ref": "#/components/schemas/GoalOverrideResponse"
            },
            "title": "Goal Overrides",
            "type": "array"
          },
          "goal_type": {
            "$ref": "#/components/schemas/GoalType",
            "default": "target"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ProjectKind",
            "default": "side_project"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "weekly_goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Goal"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ProjectResponse",
        "type": "object"
      },
      "ProjectsListItemResponse": {
        "description": "The list response shape \u2014 same fields as ProjectResponse, plus\noptional aggregations populated when GET /api/projects/ is called\nwith `include=totals,this_week,last_tracked`. Each field is None\nwhen its corresponding include flag wasn't requested. Backward\ncompatible because absent values stay None.\n\nAdded in P3.0 of the project-management revamp so the /projects\nindex page (P3.1) can load with a single round-trip instead of\nthe previous N+1 fan-out the UI did per project.",
        "properties": {
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "autostart_repos": {
            "items": {
              "type": "string"
            },
            "title": "Autostart Repos",
            "type": "array"
          },
          "balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "contract": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Contract"
              },
              {
                "type": "null"
              }
            ]
          },
          "contract_expected": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Expected"
          },
          "contract_remaining": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Remaining"
          },
          "contract_worked": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Worked"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "effective_goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Goal"
          },
          "effective_goal_overridden": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Goal Overridden"
          },
          "effective_goal_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoalType"
              },
              {
                "type": "null"
              }
            ]
          },
          "github_repo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repo"
          },
          "goal_overrides": {
            "items": {
              "$ref": "#/components/schemas/GoalOverrideResponse"
            },
            "title": "Goal Overrides",
            "type": "array"
          },
          "goal_type": {
            "$ref": "#/components/schemas/GoalType",
            "default": "target"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ProjectKind",
            "default": "side_project"
          },
          "last_tracked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tracked At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "total_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Minutes"
          },
          "weekly_goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Goal"
          },
          "weekly_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Minutes"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ProjectsListItemResponse",
        "type": "object"
      },
...
      "UpdateProjectRequest": {
        "description": "Request body for updating a project.\n\nThe update is a wholesale replace, so a client that predates a field\nwould reset it on every colour change if \"not sent\" read as \"clear\".\n`kind` and `contract` therefore tell the two apart through\n`model_fields_set`: left out, the stored value is kept; `contract: null`\nclears the contract. `kind` has no empty state, so null reads as left\nout there.",
        "properties": {
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "autostart_repos": {
            "items": {
              "type": "string"
            },
            "title": "Autostart Repos",
            "type": "array"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "contract": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Contract"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "github_repo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repo"
          },
          "goal_type": {
            "$ref": "#/components/schemas/GoalType",
            "default": "target"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "weekly_goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Goal"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "UpdateProjectRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A time tracking application",
    "title": "Beats API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/": {
      "get": {
        "description": "List projects, optionally augmented with aggregation fields.\n\nBackward compatible: without `include`, the response carries the same\nfields as ProjectResponse plus null aggregation slots \u2014 older clients\nthat ignore unknown null fields keep working.\n\nP3.0 of the project-management revamp. First cut loops per-project\nserver-side; if profiling shows it's needed, a follow-up (P3.0b)\nmoves to a Mongo $facet pipeline for true batched aggregation.",
        "operationId": "list_projects_api_projects__get",
        "parameters": [
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          },
          {
            "description": "Comma-separated aggregations to populate per project. Supported: 'totals' (total_minutes), 'this_week' (weekly_minutes + effective_goal trio, and on a day job the contract's expected / worked / remaining / balance), 'last_tracked' (last_tracked_at). Omitted \u21d2 slim response.",
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated aggregations to populate per project. Supported: 'totals' (total_minutes), 'this_week' (weekly_minutes + effective_goal trio, and on a day job the contract's expected / worked / remaining / balance), 'last_tracked' (last_tracked_at). Omitted \u21d2 slim response.",
              "title": "Include"
            }
          },
          {
            "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
            "in": "query",
            "name": "tz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
              "title": "Tz"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectsListItemResponse"
                  },
                  "title": "Response List Projects Api Projects  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Projects",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "description": "Create a new project.",
        "operationId": "create_project_api_projects__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "Projects"
        ]
      },
      "put": {
        "description": "Update an existing project.",
        "operationId": "update_project_api_projects__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/projects/{project_id}/absences": {
      "get": {
        "description": "Absences on the project within [start, end], by date. Each bound\ndefaults on its own to the current calendar year in `tz`.",
        "operationId": "list_absences_api_projects__project_id__absences_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "description": "Defaults to 1 January of this year.",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Defaults to 1 January of this year.",
              "title": "Start"
            }
          },
          {
            "description": "Defaults to 31 December of this year.",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Defaults to 31 December of this year.",
              "title": "End"
            }
          },
          {
            "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
            "in": "query",
            "name": "tz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
              "title": "Tz"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AbsenceResponse"
                  },
                  "title": "Response List Absences Api Projects  Project Id  Absences Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Absences",
        "tags": [
          "Absences"
        ]
      },
      "post": {
        "description": "Record an absence. One per date: posting again for the same date\nreplaces the first and still answers 201, since the client asked to\ncreate and gets back the absence now on record.",
        "operationId": "record_absence_api_projects__project_id__absences_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbsenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbsenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Absence",
        "tags": [
          "Absences"
        ]
      }
    },
    "/api/projects/{project_id}/absences/{absence_id}": {
      "delete": {
        "description": "Remove an absence. 404 when it is not on this project.",
        "operationId": "delete_absence_api_projects__project_id__absences__absence_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "absence_id",
            "required": true,
            "schema": {
              "title": "Absence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Absence",
        "tags": [
          "Absences"
        ]
      }
    },
...
  }
}